requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
numpy==1.26.4
pydantic==2.6.3
pydantic_core==2.16.3
pymongo==4.6.2
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.27.1
